# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's BASE",
        "description": "API's for transportation management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/base-backend/public/api/permission": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Obtener información con filtros y ordenamiento",
                "operationId": "fc1a44396875d3a2d9349b97ec27e0a9",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrar por tipo",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha de inicio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha de fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Entornos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Permission"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/base-backend/public/api/permission/{id}": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Obtener detalles de un permission por ID",
                "operationId": "d06224d9ef2ccfd009afdf08728e8735",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de Permission",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Permission"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Permission No Encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Permission No Encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Permission": {
                "title": "Permission",
                "description": "Permission model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "users"
                    },
                    "type": {
                        "type": "string",
                        "example": "Tipo-01"
                    },
                    "status": {
                        "type": "string",
                        "example": "Activo"
                    },
                    "person_id": {
                        "description": "Person Id",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}